+++ projects/pastanaga-angular/src/lib/expander/expander.component.spec.ts
@@ -1,4 +1,4 @@
-import { ExpandComponent } from './expand.component';
+import { ExpanderComponent } from './expander.component';
 import { createComponentFactory, Spectator } from '@ngneat/spectator/jest';
 import { PaButtonModule } from '../button/button.module';
 import { PaTranslateModule } from '../translate/translate.module';
@@ -8,12 +8,12 @@ import { fakeAsync, tick } from '@angular/core/testing';
 describe('ExpandComponent', () => {
     const createComponent = createComponentFactory({
         imports: [MockModule(PaButtonModule), MockModule(PaTranslateModule)],
-        component: ExpandComponent,
+        component: ExpanderComponent,
         detectChanges: false,
     });
 
-    let component: ExpandComponent;
-    let spectator: Spectator<ExpandComponent>;
+    let component: ExpanderComponent;
+    let spectator: Spectator<ExpanderComponent>;
 
     beforeEach(() => {
         spectator = createComponent();
@@ -38,21 +38,42 @@ describe('ExpandComponent', () => {
         expect(component.expanded).toBe(true);
     });
 
-    describe('ngAfterViewInit', () => {
+    describe('updateContentHeight', () => {
         it('should set contentHeight CSS variable in a timeout', fakeAsync(() => {
             const setProperty = jest.fn();
             // @ts-ignore access private member
             component.elementRef.nativeElement.style.setProperty = setProperty;
-            component.expandContent = {
+            component.expanderContent = {
                 nativeElement: {
                     getBoundingClientRect: jest.fn(() => ({ height: 100 })),
                 },
             };
 
-            component.ngAfterViewInit();
+            // @ts-ignore access private member
+            component.updateContentHeight();
             expect(setProperty).not.toHaveBeenCalled();
             tick();
             expect(setProperty).toHaveBeenCalledWith('--contentHeight', '100px');
         }));
+    });
+
+    describe('ngAfterViewInit', () => {
+        it('should call updateContentHeight', () => {
+            const updateContentHeight = jest.fn();
+            //@ts-ignore access private member
+            component.updateContentHeight = updateContentHeight;
+            component.ngAfterViewInit();
+            expect(updateContentHeight).toHaveBeenCalled();
+        });
+    });
+
+    describe('contentLoaded', () => {
+        it('should call updateContentHeight', () => {
+            const updateContentHeight = jest.fn();
+            //@ts-ignore access private member
+            component.updateContentHeight = updateContentHeight;
+            component.contentLoaded = true;
+            expect(updateContentHeight).toHaveBeenCalled();
+        });
     });
 });